server/requirements.txt
Flask
python-dotenv
python-chess
Flask-SQLAlchemy
flask_cors
flask-socketio